checkout-api/requirements.txt
Django==3.2.11
djangorestframework==3.13.1
kafka-python==2.0.2
sqlparse==0.4.2